v2.0
-Replaced JSON read/write with MongoDB
-Changed .warnings to look up own warnings
-.warnings @user restricted to mods
-filterd mod-only commands from help
-added .mod for mod commands
-refactored mute
-fixed .ranks steam name display issues
-added bot logging

+++ src/index.js
@@ -1,24 +1,18 @@
 require('dotenv').config();
-const {
-    readdirSync,
-} = require('fs');
-const {
-    join,
-} = require('path');
+const logger = require('./util/logger.js');
+const { readdirSync } = require('fs');
+const { join } = require('path');
 const AlecClient = require('./struct/Client');
-const {
-    Collection,
-} = require('discord.js');
+const { Collection } = require('discord.js');
 const client = new AlecClient({
     token: process.env.DISCORD_TOKEN,
     prefix: process.env.DISCORD_PREFIX,
 });
 const u = require('./util/utilities.js');
-const {
-    maybeCreateMemberData,
-} = require('./util/utilities.js');
-const memberdata = require('./data/memberdata.json');
-const MembersInfo = require('./struct/MembersInfo.js');
+const { getAllGuildMembers, getLastSuggestionMessageOnRestart } = require('./util/utilities.js');
+const MuteData = require('./data/models/mutedata.js');
+const MemberInfo = require('./struct/MembersInfo.js');
+const MemberData = require('./data/models/memberdata.js');
 
 const commandFiles = readdirSync(join(__dirname, 'commands')).filter(file => file.endsWith('.js'));
 for (const file of commandFiles) {
@@ -28,25 +22,32 @@ for (const file of commandFiles) {
 
 client.once('ready', () => {
     console.log(`${client.user.username} READY!`);
-    // eslint-disable-next-line no-unused-vars
-    function generateAllMembersInfo(member, memberId, allMembersInfoMap) {
-        maybeCreateMemberData(memberId);
-        Object.assign(client.memberinfo, { [memberId]: new MembersInfo(memberdata[memberId]) });
-        client.memberinfo[memberId].name = member.displayName;
-    }
+    getLastSuggestionMessageOnRestart(client);
     const guild = client.guilds.first();
-    const allMembersInfoMap = guild.members;
-    allMembersInfoMap.forEach(generateAllMembersInfo);
+    getAllGuildMembers(guild);
+    const mutedRole = guild.roles.find(role => role.name === 'muted');
+    client.setInterval(() => {
+        MuteData.find().then(mutesDb => {
+            mutesDb.forEach(mutes => {
+                const member = guild.members.get(mutes._id);
+                if (Date.now() > mutes.time) {
+                    member.removeRole(mutedRole);
+                    // eslint-disable-next-line max-nested-callbacks
+                    MuteData.findByIdAndDelete(mutes._id).then(() => console.log('deleted database entry'));
+                }
+            });
+        });
+    }, 60000);
 });
-
 
 client.on('message', message => {
     u.persistSuggestions(message);
-    if (!message.content.startsWith(client.config.prefix) || message.author.bot) return;
+    if (!message.content.startsWith(client.config.prefix) || message.author.bot || message.channel.type === 'dm') return;
     const args = message.content.slice(client.config.prefix.length).split(/ +/);
     const commandName = args.shift().toLowerCase();
     const command = client.commands.get(commandName) || client.commands.find(cmd => cmd.aliases && cmd.aliases.includes(commandName));
     if (!command) return;
+    logger.info(message, message.content);
     if (command.args && !args.length) {
         message.delete();
         let reply = `You didn't provide the correct arguments, ${message.author}!`;
@@ -79,9 +80,27 @@ client.on('message', message => {
     }
 });
 client.on('guildMemberAdd', (member) => {
-    maybeCreateMemberData(member.id);
-    Object.assign(client.memberinfo, { [member.id]: new MembersInfo(memberdata[member.id]) });
-    client.memberinfo[member.id].name = member.displayName;
+    MemberData.findById(member.id).lean().exec(function(err, foundMember) {
+        if (err) console.log(err);
+        if (!foundMember) {
+            const newMemberData = new MemberData({
+                _id: member.id,
+                userId: member.id,
+                name: member.displayName,
+                warnings: 0,
+                warningreasons: [],
+                mutecount: 0,
+                mutehistory: ['\u200B'],
+                steamid: '',
+            });
+            newMemberData.save()
+                .then(newMember => Object.assign(client.memberinfo, { [member.id]: new MemberInfo(newMember._doc) }))
+                .catch(err => console.log(err));
+        }
+        else {
+            Object.assign(client.memberinfo, { [member.id]: new MemberInfo(foundMember) });
+        }
+    });
 });
 client.on('error', (error) => {
     console.error(error);
@@ -90,7 +109,6 @@ client.on('error', (error) => {
 client.on('disconnect', (error) => {
     console.error(error);
     client.login(client.config.token);
-
 });
 
 client.on('unhandledRejection', error => console.error('Uncaught Promise Rejection', error));
